+++ changelog.md
@@ -1,4 +1,14 @@
 # changelog
+
+## 2025-04-24
+- 修改原有的评论模块为Giscus，因为原有的有不少问题，新的安装地址：https://giscus.app/ 请按照giscus官网的指导操作更换`.vitepress/theme/components/CommentGiscus.vue`中的信息
+- 默认开启评论 如果不想开启 需要在md 文件头部加入 `comment: false` , eg：`pages/tags.md`
+- 变更多文件有点多，可能需要老用户diff一下文件差异，重新初始化
+
+## 2025-04-23
+- /posts/下的文章按目录存放，而不必存放到Posts的根目录
+- 发布时排除 `trash` `private-notes` `draft` 这三个目录的md文档。
+
 ## 2025-03-17 标签 tags页面
 tag最多的内容排在最前面
 ## 2024-07-15 项目更ts化